feat!: support internal elements in multiple shows
BREAKING CHANGES: various API changes due to the Rundown no longer
being restricted to using a single Show. To support controlling elements
across multiple shows at the same time, the showId (previously also known
as showName) has to be provided for each internal element, as a property
of the InternalElementId.